Review Exception renaming and renderer removal
