--- components/ILIAS/UI/src/Component/MessageBox/Exception/LinksAndEntityListConflictException.php
@@ -1,35 +1,0 @@
-<?php
-
-/**
- * This file is part of ILIAS, a powerful learning management system
- * published by ILIAS open source e-Learning e.V.
- *
- * ILIAS is licensed with the GPL-3.0,
- * see https://www.gnu.org/licenses/gpl-3.0.en.html
- * You should have received a copy of said license along with the
- * source code, too.
- *
- * If this is not the case or you just want to try ILIAS, you'll find
- * us at:
- * https://www.ilias.de
- * https://github.com/ILIAS-eLearning
- *
- *********************************************************************/
-
-declare(strict_types=1);
-
-namespace ILIAS\UI\Component\MessageBox\Exception;
-
-use Exception;
-
-/**
- * This exception indicates that an UI component was accepted by the JF but is
- * not backed by a real implementation.
- */
-class LinksAndEntityListConflictException extends Exception
-{
-    public function __construct()
-    {
-        parent::__construct('The Message Box contains both links and entity lists. This is not allowed.');
-    }
-}
